+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 ### Changes (non-breaking)
 - Applayout navigation arrow is now optionnal by adding `.with-navigation-arrow` to the #main-navigation
 ### Bug fixes
+- `luid-on-scroll-bottom`: infinite scroll did not work on browsers like FireFox
 - `luid-user-picker`: fixed a bug where user were duplicated
 - `luid-image-picker`: fixed issue with loading states not showing
+
 
 ## 3.1.11 - [release](https://github.com/LuccaSA/lucca-ui/releases/tag/3.1.11)
 ### Changes (non-breaking)